requirements.txt
anthropic>=0.42.0
python-dotenv>=1.0.0
rich>=13.0.0
pywebview>=5.0
pystray>=0.19.0
Pillow>=10.0.0
duckduckgo-search>=7.0.0
pyperclip>=1.9.0
